Refer to Style as S for brevity, split Daily calendar into header/column/item constituents
Also add a helper so that we don't keep repeating "elm-agenda__", but
also making the in-line declaration of classes more obvious.

Still not happy with how I'm referring to styles, need to spend some
time to think about what is going to be a nice API for folks to apply
their own styles while getting nice layout out-of-the-box.