Add some safeguards to javascript in example app
It's entirely reasonable that we wouldn't be able to find a quarter that
corresponds to the supplied position, i.e. when the user drags an event
outside of the viewport, so handle this by simply disregarding the
request for a quarter at that position. Time will tell if this approach
is sufficient, or if we actually need to provide some feedback about
what happened.