Begin operating over virtualised events to allow caching rather than always persisting changes back to the consumer
This is an importance change to unlock creating events via
dragging.

Previously we persisted every change to an event back to the consumer,
so every increment in dragging an event would result in calling the
update function provided by the consumer. There are numerous problems
with this approach, including that it forced us to try and persist an
event immediately after beginning a drag.

What we really want to do is virtually create a proto event until the
user releases the drag, then prompt them for event details, and only
then persist the event back to the consumer.

This has the added benefit of somewhat simplifying the API, as the
consumer no longer needs to pass their events to the view (we already
have the virtualised events in the model) and they can pass just the
EventMapping rather than the full Config in some cases.