Implement the skeleton of basic keyboard shortcut functionality
This change introduces a settings modal with a single item: the ability
to turn on/off keyboard shortcuts.

When shortcuts are enabled, this change also turns on the display of a
new modal that shows help for keyboard shortcuts. This modal is toggled
with the / or ? keys.

There are numerous aspects of this change that are unsatisfying, however
it's the simplest thing that works, so serves as a good starting point.

In particular, we'll need to address:
* The bug mentioned in a comment above the keyboard shortcut checkbox
* The heavy use of modals, and the increasing amount of code duplication
  between these modals
* The inclusion of all setting-related fields directly on the Model
  record
* The use of KeyCode, which is deprecated in browsers
* The introduction of the NoOp Msg, which is necessary because we
  don't have a meaningful Msg to send for all key downs/ups